+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 11.2.0.1
+
+- Added support for the Retail Bank UI refresh in 11.2.0
+  - Bank Bags, Reagent Bank, and Void Storage have been removed from the game.
+  - The Warband Bank shares options with the Bank because it uses the same inventory system.
+
 # Version 11.2.0.0
 
 - Updated TOC for Pandaria Classic 5.5.0